Fix check 7 fail

+++ 0x15-file_io/1-create_file.c
@@ -9,7 +9,7 @@
  */
 int create_file(const char *filename, char *text_content)
 {
-	ssize_t i, opn, wrt;
+	int i, opn, wrt;
 
 	if (filename == NULL)
 		return (-1);
